beerjs_5/cloudfunction/requirements.txt
numpy>=1.14.1
scipy>=0.19.0
pandas>=0.22.0
python-dateutil>=2.4.2
pytz>=2015.6
six>=1.9.0
jsonschema>=2.5.1
backports.functools_lru_cache>=1.5
google-cloud-storage
pandas-gbq
google-cloud-bigquery
dill
xgboost